Atividade3/projeto-gRPC-IoT/requirements.txt
grpcio
grpcio-tools
google-api-python-client